FE: update refreshToken handling; update refreshToken interceptor

+++ Client/src/app/core/_interceptors/refresh-token.interceptor.ts
@@ -3,35 +3,45 @@ import { inject } from '@angular/core';
 import { TokenService } from '../_services/token/token.service';
 import { catchError, switchMap, throwError } from 'rxjs';
 import { ITokens } from '../_models/tokens.model';
+import { IAuthTokensResponseDto } from '../_models/DTOs/authDto.model';
+
+let isRefreshing = false;
 
 export const refreshTokenInterceptor: HttpInterceptorFn = (req, next) => {
   const tokenService = inject(TokenService);
 
   return next(req).pipe(
     catchError((error: HttpErrorResponse) => {
-      if (error.status === 401) {
+      if (error.status === 401 && !isRefreshing) {
+        isRefreshing = true;
         const accessToken = tokenService.getAccessToken();
         const refreshToken = tokenService.getRefreshToken();
 
         if (accessToken && refreshToken) {
           return tokenService.refreshToken({ accessToken, refreshToken }).pipe(
-            switchMap((newTokens: ITokens) => {
+            switchMap((newTokens: IAuthTokensResponseDto) => {
               tokenService.setAccessToken(newTokens.accessToken);
-              tokenService.setRefreshToken(newTokens.refreshToken);
+              tokenService.setRefreshToken({
+                refreshToken: newTokens.refreshToken,
+                expiresAt: newTokens.expiresAt,
+              });
 
               const clonedReq = req.clone({
                 setHeaders: {
                   Authorization: `Bearer ${newTokens.accessToken}`,
                 },
               });
+              isRefreshing = false;
               return next(clonedReq);
             }),
             catchError(refreshError => {
+              isRefreshing = false;
               tokenService.removeTokens();
               return throwError(refreshError);
             })
           );
         } else {
+          isRefreshing = false;
           tokenService.removeTokens();
           return throwError(error);
         }
